// src/shared/model/fake/fakeDataModel.ts
export class Vendormodel {
        VendorName?: string = "";
        IndustryID?: number = 0;
        VendorEmail?: string = "";
        PrefPaymentMethod?: string = "";
        KRAPinURL?: string = "";
        COIURL?: string = "";
        DirectorPinURL?: string = "";
        TOCAcceptInd?: boolean = false;
        FullName?: string = "";
        PhoneNumber?: string = "";
        UserName?: string = "";
        Password?: string = "";
        ImageURL?: string = "";
        DeviceType?: string = "";
}

export class vendorlogin {
        UserName?: string = ""
        Password?: string = ""
        DeviceType?: string = ""

}